Remove process locking and refactor singal handling
1) refactor signal handling for tracking default
revision in its own signal. This allows to remove
an unnecessary indirection and helps to improve
revision-change-handling when the deployment list
has changed.

2) remove multi-process locking..

.. as it was based on incorrect assumptions on
how concurrency is handled by OSTree.

OSTree has API to lock the sysroot, but not the
repository itself. As we can not lock the repo
and writes to the repo are atomic, reading from
the repo in general is safe. Except when the data
has been read, there is no guarantee that the data
won't disapear under us (because some other process
has done repo pruning).

Concurrent writing to the repo is handled by OSTree.
When we deploy new sysroot version via "ostree deploy",
it locks the sysroot with the sysroot locking API, so
we don't need to do any additinal locking.

Change-Id: Ide593f9bfbe93827093a6693eeb10cc3b315e000
Reviewed-by: Gatis Paeglis <gatis.paeglis@qt.io>