Expose default revision and metadata
Default system is the system that is scheduled to be
booted into during the next system startup. With the
previous 3 metadata versions it was possible to rotate
to a state where we can't access and present to the user
what currently is scheduled for the next system boot:

Booted  : b
Remote  : b
Rollback: a

call ::rollback() => [ a -> b ]

Booted   : b
Remote   : b
Rollback : b

Now if we want to present a message box to a user when
the system shutdown has been initiated, we don't know what
the system will boot into. This of course could be worked
around by saving a state in user application, but the better
solution is to expose default revision.

With this API, the previous example after the rollback will
contain:

Booted   : b
Remote   : b
Rollback : b
Default  : a

Change-Id: Id123ed8ef76f0284a02899daf22aef0716c78e4e
Reviewed-by: Gatis Paeglis <gatis.paeglis@qt.io>