Do not use PCH on qotaclientasync.cpp
qotaclientasync.cpp includes ostree.h which includes headers from glib2.
These headers also defines signals which generated compile errors due
to Qt:s signals definition.

One solution on this problem is to include the ostree headers before
including Qt-header files. But when using pre-compiled headers qmake
will automatically generate PCH for the modules dependencies, in this
case QtCore. The PCH are included using the -include argument in gcc
which includes the PCH just as doing a #include on the first line in
the source file. This means that when using PCH we will
include QtCore, which will include qobjectdefs.h, before ostree.h

By adding qotaclientasync.cpp to NO_PCH_SOURCES instead of the "normal"
SOURCES we tell qmake to not use PCH for that specific file.

Change-Id: Ideb1c6aaa98905d7f516761087b7a67bb1e3f402
Reviewed-by: Gatis Paeglis <gatis.paeglis@qt.io>